+++ meteor-packages/mdc-styleless/main.js
@@ -18,5 +18,8 @@ export * from "./bundle.js";
 export {
   meteorPackageName as name,
   version,
-  mdcVersion
 } from './package.json';
+
+export {
+  version as mdcVersion,
+} from './material-components-web.json';